requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.9.3
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
autobahn==23.1.2
autocfg==0.0.8
Automat==22.10.0
av==8.0.1
Babel==2.11.0
backcall==0.2.0
bcolz==1.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
catkin-pkg==0.5.2
catkin-tools==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudpickle==2.2.1
cmaes==0.9.1
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.2
configparser==5.3.0
constantly==15.1.0
continual-inference==1.1.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
cython-bbox==0.1.3
DCNv2 @ git+https://github.com/cidl-auth/DCNv2@a2f80597fc847c4ceff550fcd689182642c1b210
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dlib==19.24.0
docutils==0.17.1
eagerx==0.1.13
eagerx-examples==0.1.5
easydict==1.10
empy==3.3.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
filterpy==1.4.5
flake8==6.0.0
flake8-import-order==0.18.2
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gdown==4.6.2
gluoncv==0.11.0b20210908
google-auth==2.16.0
google-auth-oauthlib==0.4.6
graphviz==0.8.4
greenlet==2.0.2
grpcio==1.51.1
gym==0.20.0
h5py==2.10.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.6.0
imagesize==1.4.1
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
imutils==0.5.4
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lap==0.4.0
librosa==0.8.0
lightning-utilities==0.9.0
llvmlite==0.36.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
motmetrics==1.2.0
mpmath==1.2.1
multidict==6.0.4
munkres==1.1.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.8.8
ninja==1.11.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.53.0
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.8.0
onnxruntime==1.3.0
opencv-python==4.5.1.48
optuna==3.1.0
osrf-pycommon==2.0.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.13.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
progress==1.6
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyglet==1.5.27
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.1
pytest-benchmark==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.6
pytorch-lightning==1.9.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rosdep==0.22.1
rosdistro==0.9.0
roslibpy==1.4.1
rospkg==1.4.0
rsa==4.9
Rx==3.2.0
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soundfile==0.11.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinx-autodoc-typehints==1.19.1
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.3
stable-baselines3==1.1.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.8.10
tenacity==8.2.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==1.1.0
terminado==0.17.1
tifffile==2023.2.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.5.23
Twisted==22.10.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
vcstool==0.3.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
xmljson==0.2.1
xmltodict==0.13.0
yacs==0.1.8
yarl==1.8.2
zipp==3.13.0
zope.interface==5.5.2
